Moved flies from /external/hapi folder to /hl7 folder
