+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/external/localfile/LocalFileFetcher.java
@@ -1,7 +1,7 @@
 package gov.hhs.cdc.trustedintermediary.rse2e.external.localfile;
 
 import gov.hhs.cdc.trustedintermediary.rse2e.FileFetcher;
-import gov.hhs.cdc.trustedintermediary.rse2e.HL7FileStream;
+import gov.hhs.cdc.trustedintermediary.rse2e.hl7.HL7FileStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
