+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/hl7/HL7FileStream.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.rse2e;
+package gov.hhs.cdc.trustedintermediary.rse2e.hl7;
 
 import java.io.InputStream;
