+++ rs-e2e/src/test/groovy/gov/hhs/cdc/trustedintermediary/rse2e/AutomatedTest.groovy
@@ -2,8 +2,9 @@ package gov.hhs.cdc.trustedintermediary.rse2e
 
 import gov.hhs.cdc.trustedintermediary.context.TestApplicationContext
 import gov.hhs.cdc.trustedintermediary.rse2e.external.azure.AzureBlobFileFetcher
-import gov.hhs.cdc.trustedintermediary.rse2e.external.hapi.HapiHL7FileMatcher
-import gov.hhs.cdc.trustedintermediary.rse2e.external.hapi.HapiHL7ExpressionEvaluator
+import gov.hhs.cdc.trustedintermediary.rse2e.hl7.HL7FileStream
+import gov.hhs.cdc.trustedintermediary.rse2e.hl7.HapiHL7FileMatcher
+import gov.hhs.cdc.trustedintermediary.rse2e.hl7.HapiHL7ExpressionEvaluator
 import gov.hhs.cdc.trustedintermediary.external.jackson.Jackson
 import gov.hhs.cdc.trustedintermediary.rse2e.external.localfile.LocalFileFetcher
 import gov.hhs.cdc.trustedintermediary.wrappers.HealthDataExpressionEvaluator
