+++ rs-e2e/src/test/groovy/gov/hhs/cdc/trustedintermediary/rse2e/hl7/HL7ParserTest.groovy
@@ -1,4 +1,5 @@
-package gov.hhs.cdc.trustedintermediary.rse2e
+package gov.hhs.cdc.trustedintermediary.rse2e.hl7
+
 
 import spock.lang.Specification
