+++ rs-e2e/src/test/groovy/gov/hhs/cdc/trustedintermediary/rse2e/hl7/HapiHL7FileMatcherExceptionTest.groovy
@@ -1,4 +1,5 @@
-package gov.hhs.cdc.trustedintermediary.rse2e.external.hapi
+package gov.hhs.cdc.trustedintermediary.rse2e.hl7
+
 
 import spock.lang.Specification
